Correct paths in moved react components
