+++ client/app/bundles/HelloWorld/startup/AdminDashboardAppClient.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AdminDashboardRouter from '../containers/AdminDashboardRouter';
+import AdminDashboardRouter from '../../admin_dashboard/containers/AdminDashboardRouter';
 
 const AdminDashboardApp = (props) => {
   return(<AdminDashboardRouter {...props} />);
