+++ client/app/bundles/admin_dashboard/components/create_new_accounts.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import DropdownMenu from '../progress_reports/dropdown_filter'
+import DropdownMenu from '../../HelloWorld/components/progress_reports/dropdown_filter'
 
 export default class CreateNewAccounts extends React.Component {
   constructor(props) {
