+++ client/app/bundles/admin_dashboard/containers/AdminDashboardContainer.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import SubnavTabs from '../components/admin_dashboard/subnav_tabs';
+import SubnavTabs from '../components/subnav_tabs';
 
 const AdminDashboardContainer = (props) => {
   return (
     <div className="tab-content">
       <div className="tab-pane active" id="class-manager">
         <SubnavTabs />
-        <div id="admin-dashboard" data-id={props.id}>
+        <div id="admin-dashboard">
           {props.children}
         </div>
       </div>
